pyproject.toml
[project]
name = "mini-claw"
version = "0.1.0"
description = "Minimal local OpenClaw-style agent"
requires-python = ">=3.13.12"
dependencies = [
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "openai>=1.0",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.34",
  "sse-starlette>=2.0",
  "playwright>=1.49",
]
